+++ R/scales_cross_blended.R
@@ -1,22 +1,22 @@
-#' Cross blended Hypsometric Tints fill scales
+#' Cross blended Hypsometric Tints scales
 #'
 #' @description
 #'
 #' Implementation of the cross blended hypsometric gradients presented on
 #' \doi{10.14714/CP69.20}. The following fill scales and palettes are provided:
 #'
-#' - `scale_fill_cross_blended_d()`: For discrete values.
-#' - `scale_fill_cross_blended_c()`: For continuous values.
-#' - `scale_fill_cross_blended_b()`: For binning continuous values.
+#' - `scale_*_cross_blended_d()`: For discrete values.
+#' - `scale_*_cross_blended_c()`: For continuous values.
+#' - `scale_*_cross_blended_b()`: For binning continuous values.
 #' - `cross_blended.colors()`: A gradient color palette. See also
 #'   [grDevices::terrain.colors()] for details.
 #'
 #' An additional set of scales is provided. These scales can act as
 #' [hypsometric (or bathymetric) tints](https://en.wikipedia.org/wiki/Hypsometric_tints).
 #'
-#' - `scale_fill_cross_blended_tint_d()`: For discrete values.
-#' - `scale_fill_cross_blended_tint_c()`: For continuous values.
-#' - `scale_fill_cross_blended_tint_b()`: For binning continuous values.
+#' - `scale_*_cross_blended_tint_d()`: For discrete values.
+#' - `scale_*_cross_blended_tint_c()`: For continuous values.
+#' - `scale_*_cross_blended_tint_b()`: For binning continuous values.
 #' - `cross_blended.colors2()`: A gradient color palette. See also
 #'   [grDevices::terrain.colors()] for details.
 #'
@@ -24,7 +24,7 @@
 #'
 #' @export
 #'
-#' @name scale_fill_cross_blended
+#' @name scale_cross_blended
 #'
 #' @inheritParams ggplot2::scale_fill_viridis_b
 #' @inheritParams ggplot2::continuous_scale
@@ -48,7 +48,7 @@
 #' [terra::minmax()], [ggplot2::scale_fill_viridis_c()]
 #'
 #' @return The corresponding ggplot2 layer with the values applied to the
-#' `fill` aesthetics.
+#' `fill/colour` aesthetics.
 #'
 #' @family gradients
 #'
@@ -174,7 +174,29 @@ scale_fill_cross_blended_d <- function(palette = "cold_humid", ...,
   )
 }
 #' @export
-#' @rdname scale_fill_cross_blended
+#' @rdname scale_cross_blended
+scale_colour_cross_blended_d <- function(palette = "cold_humid", ...,
+                                         alpha = 1, direction = 1) {
+  if (alpha < 0 || alpha > 1) {
+    stop("alpha level ", alpha, " not in [0,1]")
+  }
+
+  if (!direction %in% c(-1, 1)) stop("direction must be 1 or -1")
+
+  ggplot2::discrete_scale(
+    aesthetics = "colour",
+    scale_name = "cross_blended_colour_d",
+    palette = cross_blended_pal(
+      alpha = alpha,
+      direction = direction,
+      palette = palette
+    ),
+    ...
+  )
+}
+
+#' @export
+#' @rdname scale_cross_blended
 scale_fill_cross_blended_c <- function(palette = "cold_humid", ...,
                                        alpha = 1, direction = 1,
                                        na.value = NA, guide = "colourbar") {
@@ -203,7 +225,35 @@ scale_fill_cross_blended_c <- function(palette = "cold_humid", ...,
 }
 
 #' @export
-#' @rdname scale_fill_cross_blended
+#' @rdname scale_cross_blended
+scale_colour_cross_blended_c <- function(palette = "cold_humid", ...,
+                                         alpha = 1, direction = 1,
+                                         na.value = NA, guide = "colourbar") {
+  if (alpha < 0 || alpha > 1) {
+    stop("alpha level ", alpha, " not in [0,1]")
+  }
+
+  if (!direction %in% c(-1, 1)) stop("direction must be 1 or -1")
+
+  length_pal <- nrow(extract_pal(tidyterra::cross_blended_hypsometric_tints_db,
+    palette = palette
+  ))
+
+  ggplot2::continuous_scale(
+    aesthetics = "colour",
+    scale_name = "cross_blended_colour_c",
+    scales::gradient_n_pal(cross_blended_pal(
+      alpha = alpha,
+      direction = direction,
+      palette = palette
+    )(length_pal)),
+    na.value = na.value,
+    guide = guide,
+    ...
+  )
+}
+#' @export
+#' @rdname scale_cross_blended
 scale_fill_cross_blended_b <- function(palette = "cold_humid", ...,
                                        alpha = 1, direction = 1,
                                        na.value = NA, guide = "coloursteps") {
@@ -230,7 +280,35 @@ scale_fill_cross_blended_b <- function(palette = "cold_humid", ...,
   )
 }
 #' @export
-#' @rdname scale_fill_cross_blended
+#' @rdname scale_cross_blended
+scale_colour_cross_blended_b <- function(palette = "cold_humid", ...,
+                                         alpha = 1, direction = 1,
+                                         na.value = NA, guide = "coloursteps") {
+  if (alpha < 0 || alpha > 1) {
+    stop("alpha level ", alpha, " not in [0,1]")
+  }
+
+  if (!direction %in% c(-1, 1)) stop("direction must be 1 or -1")
+
+  length_pal <- nrow(extract_pal(tidyterra::cross_blended_hypsometric_tints_db,
+    palette = palette
+  ))
+  ggplot2::binned_scale(
+    aesthetics = "colour",
+    scale_name = "cross_blended_colour_b",
+    scales::gradient_n_pal(cross_blended_pal(
+      alpha = alpha,
+      direction = direction,
+      palette = palette
+    )(length_pal)),
+    na.value = na.value,
+    guide = guide,
+    ...
+  )
+}
+
+#' @export
+#' @rdname scale_cross_blended
 #'
 #' @inheritParams wiki.colors
 #' @examples
@@ -269,10 +347,10 @@ cross_blended.colors <- function(n, palette = "cold_humid",
 }
 
 #' @export
-#' @rdname scale_fill_cross_blended
+#' @rdname scale_cross_blended
 #' @details
 #'
-#' On `scale_fill_cross_blended_tint_*` palettes, the position of the gradients
+#' On `scale_*_cross_blended_tint_*` palettes, the position of the gradients
 #' and the limits of the palette are redefined. Instead of treating the color
 #' palette as a continuous gradient, they are rescaled to act as a hypsometric
 #' tint. A rough description of these tints are:
@@ -297,7 +375,6 @@ cross_blended.colors <- function(n, palette = "cold_humid",
 #' distance between colors is different depending of the type of color.
 #'  In contrast, `cross_blended.colors()` provides an uniform gradient across
 #'  colors. See **Examples**.
-
 scale_fill_cross_blended_tint_d <- function(palette = "cold_humid", ...,
                                             alpha = 1, direction = 1) {
   if (alpha < 0 || alpha > 1) {
@@ -319,7 +396,28 @@ scale_fill_cross_blended_tint_d <- function(palette = "cold_humid", ...,
 }
 
 #' @export
-#' @rdname scale_fill_cross_blended
+#' @rdname scale_cross_blended
+scale_colour_cross_blended_tint_d <- function(palette = "cold_humid", ...,
+                                              alpha = 1, direction = 1) {
+  if (alpha < 0 || alpha > 1) {
+    stop("alpha level ", alpha, " not in [0,1]")
+  }
+
+  if (!direction %in% c(-1, 1)) stop("direction must be 1 or -1")
+
+  ggplot2::discrete_scale(
+    aesthetics = "colour",
+    scale_name = "cross_blended_tint_colour_d",
+    palette = cross_blended_pal2(
+      alpha = alpha,
+      direction = direction,
+      palette = palette
+    ),
+    ...
+  )
+}
+#' @export
+#' @rdname scale_cross_blended
 scale_fill_cross_blended_tint_c <- function(palette = "cold_humid", ...,
                                             alpha = 1, direction = 1,
                                             values = NULL,
@@ -359,7 +457,47 @@ scale_fill_cross_blended_tint_c <- function(palette = "cold_humid", ...,
 }
 
 #' @export
-#' @rdname scale_fill_cross_blended
+#' @rdname scale_cross_blended
+scale_colour_cross_blended_tint_c <- function(palette = "cold_humid", ...,
+                                              alpha = 1, direction = 1,
+                                              values = NULL,
+                                              limits = NULL,
+                                              na.value = NA,
+                                              guide = "colourbar") {
+  if (alpha < 0 || alpha > 1) {
+    stop("alpha level ", alpha, " not in [0,1]")
+  }
+
+  if (!direction %in% c(-1, 1)) stop("direction must be 1 or -1")
+
+  # Use pal limits
+  coltab <- tidyterra::cross_blended_hypsometric_tints_db
+
+  if (!palette %in% coltab$pal) {
+    stop("'palette' does not match any given palette")
+  }
+
+  hypsocol <- coltab[coltab$pal == palette, ]
+  hexcol <- as.character(hypsocol$hex)
+  if (direction == -1) hexcol <- rev(hexcol)
+  if (alpha != 1) hexcol <- ggplot2::alpha(hexcol, alpha = alpha)
+
+  if (is.null(values)) values <- hypsocol$limit
+  # Reescale
+  res <- scales::rescale(values)
+  if (is.null(limits)) limits <- range(values)
+
+  ggplot2::scale_colour_gradientn(...,
+    colors = hexcol,
+    values = res,
+    limits = limits,
+    na.value = na.value,
+    guide = guide
+  )
+}
+
+#' @export
+#' @rdname scale_cross_blended
 scale_fill_cross_blended_tint_b <- function(palette = "cold_humid", ...,
                                             alpha = 1, direction = 1,
                                             values = NULL,
@@ -400,7 +538,48 @@ scale_fill_cross_blended_tint_b <- function(palette = "cold_humid", ...,
 }
 
 #' @export
-#' @rdname scale_fill_cross_blended
+#' @rdname scale_cross_blended
+scale_colour_cross_blended_tint_b <- function(palette = "cold_humid", ...,
+                                              alpha = 1, direction = 1,
+                                              values = NULL,
+                                              limits = NULL,
+                                              na.value = NA,
+                                              guide = "coloursteps") {
+  if (alpha < 0 || alpha > 1) {
+    stop("alpha level ", alpha, " not in [0,1]")
+  }
+
+  if (!direction %in% c(-1, 1)) stop("direction must be 1 or -1")
+
+  # Use pal limits
+  coltab <- tidyterra::cross_blended_hypsometric_tints_db
+
+  if (!palette %in% coltab$pal) {
+    stop("'palette' does not match any given palette")
+  }
+
+  hypsocol <- coltab[coltab$pal == palette, ]
+  hexcol <- as.character(hypsocol$hex)
+  if (direction == -1) hexcol <- rev(hexcol)
+  if (alpha != 1) hexcol <- ggplot2::alpha(hexcol, alpha = alpha)
+
+  if (is.null(values)) values <- hypsocol$limit
+  # Reescale
+  res <- scales::rescale(values)
+  if (is.null(limits)) limits <- range(values)
+
+  ggplot2::scale_colour_stepsn(
+    ...,
+    colors = hexcol,
+    values = res,
+    limits = limits,
+    na.value = na.value,
+    guide = guide
+  )
+}
+
+#' @export
+#' @rdname scale_cross_blended
 #' @examples
 #' # Display all the cross-blended palettes on version 2
 #'
@@ -464,3 +643,36 @@ cross_blended_pal2 <- function(alpha = 1, direction = 1, palette) {
   }
   # nocov end
 }
+
+
+#' @export
+#' @rdname scale_cross_blended
+#' @usage NULL
+scale_color_cross_blended_d <- scale_colour_cross_blended_d
+
+#' @export
+#' @rdname scale_cross_blended
+#' @usage NULL
+scale_color_cross_blended_tint_d <- scale_colour_cross_blended_tint_d
+
+
+#' @export
+#' @rdname scale_cross_blended
+#' @usage NULL
+scale_color_cross_blended_c <- scale_colour_cross_blended_c
+
+#' @export
+#' @rdname scale_cross_blended
+#' @usage NULL
+scale_color_cross_blended_tint_c <- scale_colour_cross_blended_tint_c
+
+
+#' @export
+#' @rdname scale_cross_blended
+#' @usage NULL
+scale_color_cross_blended_b <- scale_colour_cross_blended_b
+
+#' @export
+#' @rdname scale_cross_blended
+#' @usage NULL
+scale_color_cross_blended_tint_b <- scale_colour_cross_blended_tint_b