+++ NEWS.md
@@ -7,6 +7,8 @@
         `fortify.SpatRaster()`, `fortify.SpatVector()`.
 -   Three additional palettes are included on `hypso.colors()`: `"artic"`,
     `"meyers"` and `"nordisk-familjebok"`.
+-   Added colour scales to all palettes: `tidyterra::scale_colour_*`. 
+ 
 
 # tidyterra 0.3.0